Add files via upload
As you can see in my code, I use two main algorithms to help me make my code more efficient and explicit. Bubble Sort is a type of algorithm that repeats certain steps through a list to sort, compare, and exchange items in that list. Binary Search is also a type of algorithm that makes your life easier when coding, binary search is sorted by an array and divides or simplifies your list until the entire array is full.